requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/4~CAP1ugDqYZ2ZVF_54thwSWnK-8L4LO5_Zcx-VcI/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.10.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
Django==4.2.26
django-filter==25.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
fonttools==4.60.1
future @ file:///AppleInternal/Library/BuildRoots/4~CAP1ugDqYZ2ZVF_54thwSWnK-8L4LO5_Zcx-VcI/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
h11==0.16.0
hnswlib==0.8.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hypothesis==6.139.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.27
langchain-cerebras==0.5.0
langchain-core==0.3.77
langchain-fused-model==0.1.1
langchain-ollama==0.3.7
langchain-openai==0.3.34
langchain-text-splitters==0.3.11
langsmith==0.4.27
macholib @ file:///AppleInternal/Library/BuildRoots/4~CAP1ugDqYZ2ZVF_54thwSWnK-8L4LO5_Zcx-VcI/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.9
matplotlib==3.9.4
nltk==3.9.1
numpy==1.26.3
ollama==0.5.4
openai==2.0.0
openpyxl==3.1.2
orjson==3.11.3
packaging==25.0
pandas==2.1.4
pillow==11.3.0
pluggy==1.6.0
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
pytest==7.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
scikit-learn==1.4.0
scipy==1.13.1
six @ file:///AppleInternal/Library/BuildRoots/4~CAP1ugDqYZ2ZVF_54thwSWnK-8L4LO5_Zcx-VcI/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.47.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tomli==2.3.0
tqdm==4.67.1
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uv==0.8.15
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
zipp==3.23.0
zstandard==0.24.0
